v2.0/requirements.txt
opencv-python
imutils
argparse
torch
torchmetrics 
numpy
pandas
wget
Pillow
seaborn
matplotlib 
transformers
torchinfo
scipy
PyWavelets
openpyxl
patool
pyunpack
tqdm
sklearn
joblib